Process transactions is real time
Previously nodes had a transaction pool to pull from. Now transctions
process immediately. If a transaction is posted to the block winner it
will be processed in real time. If it's posted to a miner that's not the
block winner that miner will forward the transction to the block winner.

Other Changes:

* Convert miner selection to an ordered list

Preiously each reveal would elect a single winning miner. Now each
reveal produces a ordered list of miners where the winner is at the top
of the list. If that miners behaves maliciously in any way their stake
will be slashed and the next miner in the list will be elected winner.

* Refactoring